backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
python-multipart==0.0.6
tensorflow==2.18.1
numpy==1.24.3
Pillow==9.5.0
scikit-learn==1.2.2
locust==2.17.0
aiohttp==3.9.1
matplotlib==3.7.1
requests==2.31.0